back/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-mail-templated==2.6.5
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
packaging==20.9
phonenumbers==8.12.25
psycopg2==2.9.1
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
PySocks==1.7.1
python-docx==0.8.11
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0